Address comments #3
