+++ flink-connector-jdbc/src/main/java/org/apache/flink/connector/jdbc/internal/converter/PostgresRowConverter.java
@@ -18,23 +18,15 @@
 
 package org.apache.flink.connector.jdbc.internal.converter;
 
-import org.apache.flink.connector.jdbc.converter.AbstractJdbcRowConverter;
-import org.apache.flink.table.data.GenericArrayData;
-import org.apache.flink.table.types.logical.ArrayType;
-import org.apache.flink.table.types.logical.LogicalType;
-import org.apache.flink.table.types.logical.LogicalTypeRoot;
 import org.apache.flink.table.types.logical.RowType;
-import org.apache.flink.table.types.logical.utils.LogicalTypeUtils;
 
 import org.postgresql.jdbc.PgArray;
-
-import java.lang.reflect.Array;
 
 /**
  * Runtime converter that responsible to convert between JDBC object and Flink internal object for
  * PostgreSQL.
  */
-public class PostgresRowConverter extends AbstractJdbcRowConverter {
+public class PostgresRowConverter extends AbstractPostgresCompatibleRowConverter<PgArray> {
 
     private static final long serialVersionUID = 1L;
 
@@ -45,57 +37,5 @@ public class PostgresRowConverter extends AbstractJdbcRowConverter {
 
     public PostgresRowConverter(RowType rowType) {
         super(rowType);
-    }
-
-    @Override
-    public JdbcDeserializationConverter createInternalConverter(LogicalType type) {
-        LogicalTypeRoot root = type.getTypeRoot();
-
-        if (root == LogicalTypeRoot.ARRAY) {
-            ArrayType arrayType = (ArrayType) type;
-            return createPostgresArrayConverter(arrayType);
-        } else {
-            return createPrimitiveConverter(type);
-        }
-    }
-
-    @Override
-    protected JdbcSerializationConverter createNullableExternalConverter(LogicalType type) {
-        LogicalTypeRoot root = type.getTypeRoot();
-        if (root == LogicalTypeRoot.ARRAY) {
-            // note:Writing ARRAY type is not yet supported by PostgreSQL dialect now.
-            return (val, index, statement) -> {
-                throw new IllegalStateException(
-                        String.format(
-                                "Writing ARRAY type is not yet supported in JDBC:%s.",
-                                converterName()));
-            };
-        } else {
-            return super.createNullableExternalConverter(type);
-        }
-    }
-
-    protected JdbcDeserializationConverter createPostgresArrayConverter(ArrayType arrayType) {
-        // Since PGJDBC 42.2.15 (https://github.com/pgjdbc/pgjdbc/pull/1194) bytea[] is wrapped in
-        // primitive byte arrays
-        final Class<?> elementClass =
-                LogicalTypeUtils.toInternalConversionClass(arrayType.getElementType());
-        final JdbcDeserializationConverter elementConverter =
-                createNullableInternalConverter(arrayType.getElementType());
-        return val -> {
-            PgArray pgArray = (PgArray) val;
-            Object[] in = (Object[]) pgArray.getArray();
-            final Object[] array = (Object[]) Array.newInstance(elementClass, in.length);
-            for (int i = 0; i < in.length; i++) {
-                array[i] = elementConverter.deserialize(in[i]);
-            }
-            return new GenericArrayData(array);
-        };
-    }
-
-    // Have its own method so that Postgres can support primitives that super class doesn't support
-    // in the future
-    private JdbcDeserializationConverter createPrimitiveConverter(LogicalType type) {
-        return super.createInternalConverter(type);
     }
 }
